Move sync/chan into comms package
