+++ cap/csp/server/comm/chan.go
@@ -1,4 +1,4 @@
-package csp_server
+package comm
 
 import (
 	"context"
