+++ cap/csp/server/comm/sync.go
@@ -1,4 +1,4 @@
-package csp_server
+package comm
 
 import (
 	"container/list"
