+++ cap/csp/server/comm/sync_test.go
@@ -1,4 +1,4 @@
-package csp_server_test
+package comm
 
 // import (
 // 	"context"
